Refactor VariantCache implementations.
